Integrated DearIMGUI

+++ GamePlusEngine0/SpriteBatch.cpp
@@ -268,7 +268,7 @@ namespace IceEngine
 		{
 			for (int x = 0; x < 5; x++)
 			{
-				buffer = CreateQuad(buffer, x, y, (x + y) % 2);
+				buffer = CreateQuad(buffer, static_cast<float>(x), static_cast<float>(y), static_cast<float>((x + y) % 2));
 			}
 		}
